requirements.txt
streamlit
transformers
torch
Pillow
tqdm
numpy
faiss-cpu
pathlib